Move options components into web-client
