--- options/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "options",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
-    "lint": "eslint .",
-    "preview": "vite preview",
-    "watch": "vite build --watch"
-  },
-  "dependencies": {
-    "bootstrap": "^5.3.7",
-    "bootswatch": "^5.3.7",
-    "react": "^19.0.0",
-    "react-bootstrap": "^2.10.10",
-    "react-dom": "^19.0.0",
-    "react-icons": "^5.5.0",
-    "tsc-watch": "^6.2.1"
-  },
-  "devDependencies": {
-    "@eslint/js": "^9.21.0",
-    "@types/chrome": "^0.0.313",
-    "@types/react": "^19.0.10",
-    "@types/react-dom": "^19.0.4",
-    "@vitejs/plugin-react": "^4.3.4",
-    "eslint": "^9.21.0",
-    "eslint-plugin-react-hooks": "^5.1.0",
-    "eslint-plugin-react-refresh": "^0.4.19",
-    "globals": "^15.15.0",
-    "typescript": "~5.7.2",
-    "typescript-eslint": "^8.24.1",
-    "vite": "^6.3.5"
-  }
-}
