--- options/src/App.css
@@ -1,19 +1,0 @@
-#root > div {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-}
-
-.tab-content {
-    flex: 1 1 auto;
-    overflow-y: auto;
-}
-
-
-.sticky-footer {
-    position: sticky;
-    bottom: 0;
-    background-color: var(--bs-body-bg);
-    border-top: 1px solid var(--bs-border-color);
-    padding: 0.5rem 1rem;
-}
