--- options/src/index.css
@@ -1,15 +1,0 @@
-@import 'bootswatch/dist/darkly/bootstrap.min.css';
-
-html {
-    font-size: 85%;
-}
-
-.w-32 { width: 8rem; }
-.w-40 { width: 10rem; }
-
-@media (max-width: 576px) {
-    .w-32,
-    .w-40 {
-        width: 100% !important;
-    }
-}
