--- options/vite.config.ts
@@ -1,12 +1,0 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-
-// https://vite.dev/config/
-export default defineConfig({
-  plugins: [react()],
-  base: '',
-  build: {
-    outDir: '../extension/options',
-    emptyOutDir: true,
-  }
-})
