+++ package.json
@@ -6,7 +6,6 @@
   "workspaces": [
     "client",
     "map",
-    "options",
     "web-client"
   ],
   "devDependencies": {
