+++ sandbox/tsconfig.app.json
@@ -26,9 +26,6 @@
       ],
       "@map/*": [
         "../map/*"
-      ],
-      "@options/*": [
-        "../options/*"
       ]
     }
   },
