+++ sandbox/tsconfig.node.json
@@ -23,9 +23,6 @@
     ],
     "@map/*": [
       "../map/*"
-    ],
-    "@options/*": [
-      "../options/*"
     ]
   },
   "include": [
