+++ web-client/tsconfig.app.json
@@ -19,15 +19,11 @@
     "jsx": "react-jsx",
     "typeRoots": [
       "../node_modules/@types",
-      "../options/src",
       "../client/src/types",
       "./src/types",
       "node_modules/@types"
     ],
     "paths": {
-      "@options/*": [
-        "../options/*"
-      ],
       "@client/*": [
         "../client/*"
       ],
