+++ web-client/tsconfig.json
@@ -12,8 +12,7 @@
     "typeRoots": [
       "node_modules/@types",
       "./src/types",
-      "../client/src/types",
-      "../options/src"
+      "../client/src/types"
     ]
   }
 }
