+++ web-client/tsconfig.node.json
@@ -15,16 +15,12 @@
     "noEmit": true,
     "typeRoots": [
       "../node_modules/@types",
-      "../options/src",
       "../client/src/types",
       "./src/types",
       "node_modules/@types"
     ],
   },
   "paths": {
-    "@options/*": [
-      "../options/*"
-    ],
     "@client/*": [
       "../client/*"
     ],
